Add files via upload
Excessive amounts of acoustic noise pose a health risk to a majority of the population and this is the most signicant single cause of hearing loss in Australia. In addition, there are many situations where acoustic noises are annoying or at least undesirable, such as automobile cabin noise, babble noise in a restaurant, noise inside an airplane, or noise from machinery and industry plants. Thus, for a long time, there has been a strong demand for finding ways of eliminating acoustic noises to improve our lives.